Solution to Exploratory Data Analysis Course Project 1.
Plots can be found in the the upper level directory of this repository.
All plots are PNG files per specification.

The R code to create each of the plots can be found in the upper level
directory of this repository.

In addition to code generating each plot, there is a script to generate
the usable data set called get_tidy_data.R. This script downloads,
unzips, reads, cleanses, and filters the data so the plots can be
reproduced at anytime.